Separated skills/inventory/etc. under parallel scene in /ui/dashboard.js
